dynamically load methods to prevent import errors
